requirements.txt
torch==1.1.0
torchvision==0.3.0
numpy==1.16.4
scikit-learn==0.21.1
scipy==1.2.1
pillow==6.2.1